requirements.txt
aiocontextvars==0.2.2
certifi==2020.12.5
chardet==4.0.0
contextvars==2.4
grpcio==1.36.1
grpcio-tools==1.26.0
idna==2.10
immutables==0.15
loguru==0.5.3
passlib==1.7.4
peewee==3.14.4
protobuf==3.15.6
PyMySQL==1.0.2
python-consul2==0.1.5
redis==3.5.3
requests==2.25.1
rocketmq-client-python==2.0.0
six==1.15.0
urllib3==1.26.4
